refactor(context): move files
