+++ packages/core/src/utils/index.ts
@@ -1,7 +1,7 @@
 export * from "./types"
 export * from "./args"
 export * from "./middleware"
-export * from "./context"
+export * from "../contexts/context"
 export * from "./object"
 export * from "./parse-resolving-fields"
 export * from "./string"
